chore(ci): split the Edit Asset form into separate components
