+++ static/src/components/edit-asset-modal/name-field.jsx
@@ -1,0 +1,17 @@
+export const NameField = ({ formData, handleInputChange }) => {
+  return (
+    <div className="form-group row name">
+      <label className="col-4 col-form-label">Name</label>
+      <div className="col-7">
+        <input
+          className="form-control shadow-none"
+          name="name"
+          placeholder="Nickname for this asset"
+          type="text"
+          value={formData.name}
+          onChange={handleInputChange}
+        />
+      </div>
+    </div>
+  )
+}
